server/requirements.txt
absl-py==0.12.0
asgiref==3.2.10
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-mongo-sessions==0.0.4
djangorestframework==3.12.1
djongo==1.3.3
dnspython==2.0.0
dweepy==0.3.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
Keras-Preprocessing==1.1.2
keyboard==0.13.5
Markdown==3.3.4
mega.py==1.0.8
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pathlib==1.0.1
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.8
pymongo==3.11.0
pynput==1.7.2
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.2.4
tenacity==5.1.5
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.4.1
tensorflow-estimator==2.4.0
tensorflow-hub==0.7.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
